backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
pydantic==2.5.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
